mcpg/pyproject.toml
[project]
name = "kpmcpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "jsonschema>=4.25.1",
    "msal>=1.34.0",
    "pydantic>=2.12.0",
    "pymongo<4.0",
    "python-dotenv>=1.1.1",
]
